+++ lib/src/_models/_index.dart
@@ -3,6 +3,6 @@
 // All rights reserved
 
 // List all exports below in alphabetical order.
-export 'document.dart';
+// export 'document.dart';
 export 'postings.dart';
 export 'dictionary.dart';